+++ R/unark.R
@@ -13,6 +13,10 @@
 #' @param tablenames vector of tablenames to be used for corresponding files.
 #' By default, tables will be named using lowercase names from file basename with
 #' special characters replaced with underscores (for SQL compatibility).
+#' @param try_native logical, default TRUE. Should we try to use a native bulk
+#' import method for the database connection?  This can substantially speed up
+#' read times and will fall back on the DBI method for any table that fails
+#' to import.  Currently only MonetDBLite connections support this.
 #' @param ... additional arguments to `streamable_table$read` method.
 #' @details `unark` will read in a files in chunks and 
 #' write them into a database.  This is essential for processing
@@ -37,7 +41,7 @@
 #' ark(db, dir)
 #' 
 #' ## list all files in archive (full paths)
-#' files <- list.files(dir, "[.]tsv\\.bz2$", full.names = TRUE)
+#' files <- list.files(dir, "bz2$", full.names = TRUE)
 #' 
 #' ## Read archived files into a new database (another sqlite in this case)
 #' new_db <- src_sqlite(file.path(dir, "local.sqlite"), create=TRUE)
@@ -55,9 +59,9 @@ unark <- function(files,
                   overwrite = "ask",
                   encoding = Sys.getenv("encoding", "UTF-8"),
                   tablenames = NULL,
+                  try_native = TRUE,
                   ...){
   
-  assert_files_exist(files)
   assert_dbi(db_con)
   
   ## Guess streamable table
@@ -83,8 +87,10 @@ unark <- function(files,
                       overwrite = overwrite,
                       encoding = encoding,
                       tablename = tablenames[[i]],
+                      try_native = try_native,
                       ...)
            })
+  
   invisible(db_con)  
 }
 
@@ -99,7 +105,6 @@ normalize_con <- function(db_con){
 }
 
 #' @importFrom DBI dbWriteTable
-#' @importFrom progress progress_bar
 unark_file <- function(filename,
                        db_con,
                        streamable_table,
@@ -107,6 +112,7 @@ unark_file <- function(filename,
                        overwrite,
                        encoding,
                        tablename = base_name(filename),
+                       try_native = try_native,
                        ...){
     
 
@@ -114,59 +120,41 @@ unark_file <- function(filename,
     return(NULL)
   }
     
-  
-  
-  con <- compressed_file(filename, "r", encoding = encoding)
-  on.exit(close(con))
+  ## Check for a bulk importer first
+  ## FIXME use S3 method
+  bulk <- bulk_importer(db_con, streamable_table)
+  if(!is.null(bulk) && try_native){
+    t0 <- Sys.time()
+    message(paste("Native bulk importer found, attempting fast import of", basename(filename)))
+    status <- 
+      tryCatch(bulk(db_con, filename, tablename, ...),
+               error = function(e) 1)
+    if(status == 0){
+      message(sprintf("\t...Done! (in %s)", format(Sys.time() - t0)))
+      return(invisible(db_con))
+    } else {
+      message("Native import failed, falling back on R-based parser")
+    }
+  }
+
   
   ## Handle case of `col_names != TRUE`?
   ## readr method needs UTF-8 encoding for these newlines to be newlines
-  header <- read_lines(con, n = 1L, encoding = encoding)
-  if(length(header) == 0){ # empty file, would throw error
-    return(invisible(db_con))
-  }
-  reader <- read_chunked(con, lines, encoding)
-  
-  # May throw an error if we need to read more than 'total' chunks?
-  p <- progress::progress_bar$new("[:spin] chunk :current", total = 100000)
-  message(sprintf("Importing %s in %d line chunks:",
-                  basename(filename), lines))
-  t0 <- Sys.time()
-  repeat {
-    d <- reader()
-    body <- paste0(c(header, d$data), "\n", collapse = "")
-    p$tick()
-    chunk <- streamable_table$read(body, ...)
+
+  dbi_writer <- function(chunk){
     DBI::dbWriteTable(db_con, tablename, chunk, append=TRUE)
-    
-    if (d$complete) {
-      break
-    }
   }
-  message(sprintf("\t...Done! (in %s)", format(Sys.time() - t0)))
+  process_chunks(filename, 
+                 process_fn = dbi_writer,
+                 streamable_table = streamable_table,
+                 lines = lines, 
+                 encoding = encoding,
+                 ...
+                 )
   
   invisible(db_con)
 }
 
-
-# Adapted from @richfitz, MIT licensed
-# https://github.com/vimc/montagu-r
-# /blob/4fe82fd29992635b30e637d5412312b0c5e3e38f/R/util.R#L48-L60
-
-read_chunked <- function(con, n, encoding) {
-  assert_connection(con)
-  next_chunk <- read_lines(con, n, encoding = encoding)
-  if (length(next_chunk) == 0L) {
-    warning("connection has already been completely read")
-    return(function() list(data = character(0), complete = TRUE))
-  }
-  function() {
-    data <- next_chunk
-    next_chunk <<- read_lines(con, n, encoding = encoding)
-    complete <- length(next_chunk) == 0L
-    list(data = data, complete = complete)
-  }
-}
 
 
 ## Do repeatedly to remove compression extension and file extension
@@ -181,27 +169,7 @@ base_name <- function(filename){
   tolower(path)
 }
 
-#' @importFrom tools file_ext
-compressed_file <- function(path, ...){
-  con <- switch(tools::file_ext(path),
-         gz = gzfile(path, ...),
-         bz2 = bzfile(path, ...),
-         xz = xzfile(path, ...),
-         zip = unz(path, ...),
-         file(path, ...))
-}
 
-
-read_lines <- function(con,
-                       n,
-                       encoding = "unknown",
-                       warn = FALSE){
-  out <- readLines(con,
-                   n = n,
-                   encoding = encoding,
-                   warn = FALSE)
-
-}
 
 guess_stream <- function(x){  
   ext <- tools::file_ext(x)