+++ inst/doc/arkdb.R
@@ -1,14 +1,14 @@
-## ----setup, include = FALSE----------------------------------------------
+## ----setup, include = FALSE---------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----gh-installation, eval = FALSE---------------------------------------
+## ----gh-installation, eval = FALSE--------------------------------------------
 #  # install.packages("devtools")
 #  devtools::install_github("cboettig/arkdb")
 
-## ----message = FALSE-----------------------------------------------------
+## ----message = FALSE----------------------------------------------------------
 library(arkdb)
 
 # additional libraries just for this demo
@@ -17,17 +17,17 @@ library(dplyr)
 library(nycflights13)
 library(fs)
 
-## ----example-------------------------------------------------------------
+## ----example------------------------------------------------------------------
 tmp <- tempdir() # Or can be your working directory, "."
 db <- dbplyr::nycflights13_sqlite(tmp)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 dir <- fs::dir_create(fs::path(tmp, "nycflights"))
 ark(db, dir, lines = 50000)
 
 
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 fs::dir_info(dir) %>% 
   select(path, size) %>%
   mutate(path = fs::path_file(path))
@@ -37,37 +37,37 @@ fs::file_info(fs::path(tmp,"nycflights13.sqlite")) %>%
 
 
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 files <- fs::dir_ls(dir, glob = "*.tsv.bz2")
-new_db <- src_sqlite(fs::path(tmp, "local.sqlite"), create=TRUE)
+new_db <- DBI::dbConnect(RSQLite::SQLite(), fs::path(tmp, "local.sqlite"))
 
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 unark(files, new_db, lines = 50000)  
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 tbl(new_db, "flights")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 # Remove example files we created.
-DBI::dbDisconnect(new_db$con)
+DBI::dbDisconnect(new_db)
 unlink(dir, TRUE)
 unlink(fs::path(tmp, "local.sqlite"))
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 dir <- fs::dir_create(fs::path(tmp, "nycflights"))
 
 ark(db, dir, 
     streamable_table = streamable_base_csv())
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 files <- fs::dir_ls(dir, glob = "*.csv.bz2")
-new_db <- src_sqlite(fs::path(tmp,"local.sqlite"), create=TRUE)
+new_db <- DBI::dbConnect(RSQLite::SQLite(), fs::path(tmp, "local.sqlite"))
 
 unark(files, new_db,
       streamable_table = streamable_base_csv())
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 stream <- 
    streamable_table(
      function(file, ...) readr::read_tsv(file, ...),
@@ -76,12 +76,12 @@ stream <-
      "tsv")
 
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 ark(db, dir, 
     streamable_table = stream)
 
-## ----include=FALSE-------------------------------------------------------
+## ----include=FALSE------------------------------------------------------------
 DBI::dbDisconnect(db$con)
-DBI::dbDisconnect(new_db$con)
+DBI::dbDisconnect(new_db)
 unlink(dir, TRUE)