+++ tests/testthat/setup-test_db.R
@@ -1,0 +1,14 @@
+
+## Optionally: Force a backend type, otherwise will use the best available
+#Sys.setenv(ARKDB_DRIVER="MonetDBLite")
+#Sys.setenv(ARKDB_DRIVER="RSQLite")
+#Sys.setenv(ARKDB_DRIVER="duckdb")
+
+library(arkdb)
+
+## All tests only write to tempdir
+test_db <- file.path(tempdir(), "arkdb")
+Sys.setenv(ARKDB_HOME=test_db)
+
+print(paste("Testing using backend", class(local_db())))
+